windows: enable json-c and fix OCP internal-log telemetry handling (#76)
## Summary
Final update for the Windows json-c + OCP telemetry enablement branch.

This PR is now aligned on top of `mingw` and includes both the
Windows-specific telemetry/runtime fixes and the doc regeneration that
came with the branch sync.

## Scope Snapshot
- Diff vs `mingw`: **276 files changed**, **438 insertions**, **304
deletions**.
- Documentation refresh: **259 files** under `Documentation/*.html`.
- Non-doc code changes: **17 files** across `libnvme`, `plugins`, and
core CLI.

## Key Functional Changes
### Windows/json-c build and plugin wiring
- Enable json-c resolution in Windows build path.
- Ensure plugin gating/build behavior is compatible with Windows +
json-c.

### OCP telemetry path hardening
- Fix `ocp internal-log` flow and related OCP JSON/log handling paths.
- Ensure binary telemetry outputs are written in binary mode
(`O_BINARY`) to avoid CRLF corruption on Windows.
- Keep telemetry artifacts parsable/consistent
(`telemetry-log-telemetry.bin`, `telemetry-log-string.bin`,
`telemetry-log.json`).

### Platform and output safety improvements
- Add `nvme_init()` startup hook for platform init.
- Set stdout/stderr to binary mode on Windows at startup.
- Use binary-safe output path (`fwrite`) for raw output.
- Normalize socket/platform header usage in JSON print path.

### Windows libnvme compatibility/link fixes
- Update platform abstraction headers and Windows platform helpers.
- Add/adjust missing Windows stubs in `windows-stubs.c` for APIs
referenced by CLI/topology code (including accessor-related and
Linux-only helper entry points needed for linking).
- Keep accessor type include portable via platform types.

### Vendor plugin updates
- OCP plugin fixes in multiple files for build/runtime compatibility.
- Micron telemetry update to support Data Area 4 parsing path.

## Validation
- Reconfigured static build with:
  - `-Djson-c=enabled`
  - `-Ddefault_library=static`
  - `-Dwrap_mode=forcefallback`
- Full build on current `mingw-json` head succeeds (`meson compile -C
.build`, exit code 0).
- `nvme.exe` and test binaries link successfully on this branch.

## Notes
- This PR now contains substantial generated documentation updates from
branch sync/release doc regeneration in addition to the Windows
telemetry work.

---------

+++ plugins/solidigm/solidigm-smart.c
@@ -148,12 +148,12 @@ static void smart_log_item_print(struct nvme_additional_smart_log_item *item)
 			le32_to_cpu(item->thermal_throttle.count));
 		return;
 	case 0xF3:
-		printf("gain0: %u, loss0: %u, gain1: %u, loss1: %u, legacy:%lu\n",
+		printf("gain0: %u, loss0: %u, gain1: %u, loss1: %u, legacy:%llu\n",
 			le16_to_cpu(pll_item->gainCount0),
 			le16_to_cpu(pll_item->lossCount0),
 			le16_to_cpu(pll_item->gainCount1),
 			le16_to_cpu(pll_item->lossCount1),
-			int48_to_long(item->raw));
+			(unsigned long long)int48_to_long(item->raw));
 		return;
 	default:
 		printf("%"PRIu64"\n", int48_to_long(item->raw));
